+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Version 72
 *in development*
 
 * 🔨 Do not show copied to clipboard notification if Android already does.
+* 🔨 Seasons: do not say all episodes are watched if all are skipped.
 
 #### 72.0.6 🧪
 *2024-03-15*